revno: 17810 message:   Using the new user solution in user module and web api. Remvoed obsolete methods.
committer: Lars Helge Overland <larshelge@gmail.com>
branch nick: dhis2
timestamp: Sun 2014-12-28 15:15:49 +0100

+++ dhis-2/dhis-api/src/main/java/org/hisp/dhis/user/UserStore.java
@@ -72,5 +72,5 @@ public interface UserStore
      * @param params the user query parameters.
      * @return number of users.
      */
-    long getUserCount( UserQueryParams params );
+    int getUserCount( UserQueryParams params );
 }
